Refactor to domain-driven architecture for improved maintainability
- Separate domain logic from infrastructure concerns
- Create clean boundaries between business rules and external systems
- Move war state management to internal/domain/war/
- Move attack processing to internal/domain/attack/
- Move travel calculations to internal/domain/travel/
- Create application services layer for orchestration
- All domain tests passing independently
- 45% API usage reduction maintained through optimized state management
