+++ internal/domain/attack/attack_service.go
@@ -1,4 +1,4 @@
-package processing
+package attack
 
 import (
 	"time"
