+++ internal/domain/attack/attack_service_property_test.go
@@ -1,4 +1,4 @@
-package processing
+package attack
 
 import (
 	"reflect"
