+++ internal/domain/attack/attack_service_test.go
@@ -1,4 +1,4 @@
-package processing
+package attack
 
 import (
 	"testing"
