+++ internal/domain/travel/location_service.go
@@ -1,4 +1,4 @@
-package processing
+package travel
 
 import (
 	"strings"
