+++ internal/domain/travel/location_service_test.go
@@ -1,4 +1,4 @@
-package processing
+package travel
 
 import (
 	"testing"
