+++ internal/domain/travel/travel_time_property_test.go
@@ -1,4 +1,4 @@
-package processing
+package travel
 
 import (
 	"fmt"
