+++ internal/domain/travel/travel_time_service_test.go
@@ -1,4 +1,4 @@
-package processing
+package travel
 
 import (
 	"context"
