+++ internal/domain/war/war_state_manager_test.go
@@ -1,4 +1,4 @@
-package processing
+package war
 
 import (
 	"testing"
