+++ internal/processing/interface_compliance.go
@@ -1,6 +1,8 @@
 package processing
 
 import (
+	"torn_rw_stats/internal/domain/attack"
+	"torn_rw_stats/internal/domain/travel"
 	"torn_rw_stats/internal/sheets"
 	"torn_rw_stats/internal/torn"
 )
@@ -11,8 +13,8 @@ import (
 var (
 	_ TornClientInterface              = (*torn.Client)(nil)
 	_ SheetsClientInterface            = (*sheets.Client)(nil)
-	_ LocationServiceInterface         = (*LocationService)(nil)
-	_ TravelTimeServiceInterface       = (*TravelTimeService)(nil)
-	_ AttackProcessingServiceInterface = (*AttackProcessingService)(nil)
-	_ WarSummaryServiceInterface       = (*WarSummaryService)(nil)
+	_ LocationServiceInterface         = (*travel.LocationService)(nil)
+	_ TravelTimeServiceInterface       = (*travel.TravelTimeService)(nil)
+	_ AttackProcessingServiceInterface = (*attack.AttackProcessingService)(nil)
+	// WarSummaryServiceInterface compliance check moved to application/services package
 )
